Upgrade system
Another major core system down, it may be a bit large but I cannot do anything at the moment to it.
I added faction functionality changing the visuals, colors, everything depending on the faction of the tower, I also changed the faction colors to better suit the visuals I made for it.