New towers
- Universal tower: Minigun, Attacks very fast with medium damage, but overheats if left for some time or shoots for  a prolonged amount of time.
It is one of the towers that have a hidden upgrade by simply upgrading the tower, letting the tower be able to shoot more before it overheats.

- Universal tower: Anti-Immunity V1,
A weaker variant of the Anti-Immunity, it removes the immunity off of enemies but without any passive damage.
The amount of time to remove the immunity at max level is also slower by 2 seconds